api/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.2.1
python-dotenv==1.0.0
# Remove aiohttp and use only synchronous requests
# aiohttp==3.8.6
werkzeug==2.3.7
email_validator==2.2.0
psycopg2-binary==2.9.9
